Revert "board-n1: Code Cleanup fix all warnings"
This reverts commit 51fe2d0cfda9403933a19adbf6e31d962377b3ca.

+++ include/media/s5k6aafx.h
@@ -184,7 +184,7 @@ typedef enum {
 
 
 struct s5k6aafx_reg_8 {
-	u16 addr;
+	u8 addr;
 	u8 val;
 };
 struct s5k6aafx_reg {
